requirements.txt
torch
torchvision
pillow
matplotlib
tensorboard
tensorboardX
lmdb
tfrecord
addict
pyaml
opencv-python
